requirements.txt
ffmpeg-python
pyrogram
pytgcalls[pyrogram]
tgcrypto
aiohttp
idna
chardet
